JBR-4787 Rounded corners for native Window on Windows 11
