+++ src/java.desktop/windows/native/libawt/windows/awt_Window.cpp
@@ -46,6 +46,7 @@
 #include "sun_awt_windows_WCanvasPeer.h"
 
 #include <windowsx.h>
+#include <dwmapi.h>
 #include <math.h>
 #if !defined(__int3264)
 typedef __int32 LONG_PTR;
@@ -130,6 +131,11 @@ struct OpacityStruct {
 struct OpaqueStruct {
     jobject window;
     jboolean isOpaque;
+};
+// struct for _SetRoundedCorners() method
+struct RoundedCornersStruct {
+    jobject window;
+    DWM_WINDOW_CORNER_PREFERENCE type;
 };
 // struct for _UpdateWindow() method
 struct UpdateWindowStruct {
@@ -3354,6 +3360,23 @@ void AwtWindow::_SetOpaque(void* param)
     delete os;
 }
 
+void AwtWindow::_SetRoundedCorners(void *param) {
+    JNIEnv *env = (JNIEnv *)JNU_GetEnv(jvm, JNI_VERSION_1_2);
+
+    RoundedCornersStruct *rcs = (RoundedCornersStruct *)param;
+    jobject self = rcs->window;
+
+    PDATA pData;
+    JNI_CHECK_PEER_GOTO(self, ret);
+    AwtWindow *window = (AwtWindow *)pData;
+
+    DwmSetWindowAttribute(window->GetHWnd(), DWMWA_WINDOW_CORNER_PREFERENCE, &rcs->type, sizeof(DWM_WINDOW_CORNER_PREFERENCE));
+
+  ret:
+    env->DeleteGlobalRef(self);
+    delete rcs;
+}
+
 void AwtWindow::_UpdateWindow(void* param)
 {
     JNIEnv *env = (JNIEnv *)JNU_GetEnv(jvm, JNI_VERSION_1_2);
@@ -4110,6 +4133,26 @@ Java_sun_awt_windows_WWindowPeer_repositionSecurityWarning(JNIEnv *env,
     AwtToolkit::GetInstance().InvokeFunction(
             AwtWindow::_RepositionSecurityWarning, rsws);
     // global refs and mds are deleted in _RepositionSecurityWarning
+
+    CATCH_BAD_ALLOC;
+}
+
+/*
+ * Class:     sun_awt_windows_WWindowPeer
+ * Method:    setRoundedCorners
+ * Signature: (I)V
+ */
+JNIEXPORT void JNICALL
+Java_sun_awt_windows_WWindowPeer_setRoundedCorners(JNIEnv *env, jobject self, jint type)
+{
+    TRY;
+
+    RoundedCornersStruct *rcs = new RoundedCornersStruct;
+    rcs->window = env->NewGlobalRef(self);
+    rcs->type = (DWM_WINDOW_CORNER_PREFERENCE)type;
+
+    AwtToolkit::GetInstance().SyncCall(AwtWindow::_SetRoundedCorners, rcs);
+    // global refs and rcs are deleted in _SetRoundedCorners
 
     CATCH_BAD_ALLOC;
 }
