+++ src/java.desktop/windows/native/libawt/windows/awt_Window.h
@@ -249,6 +249,7 @@ public:
     static void _SetFullScreenExclusiveModeState(void* param);
     static void _GetNativeWindowSize(void* param);
     static void _OverrideHandle(void *param);
+    static void _SetRoundedCorners(void* param);
 
     inline static BOOL IsResizing() {
         return sm_resizing;
