+++ src/jetbrains.api/version.properties
@@ -6,9 +6,9 @@
 # 2. When only new API is added, or some existing API was @Deprecated - increment MINOR, reset PATCH to 0
 # 3. For major backwards incompatible API changes - increment MAJOR, reset MINOR and PATCH to 0
 
-VERSION = 1.0.0
+VERSION = 1.1.0
 
 # Hash is used to track changes to jetbrains.api, so you would not forget to update version when needed.
 # When you make any changes, "make jbr-api" will fail and ask you to update hash and version number here.
 
-HASH = B85A399915766F5414E8CA8BA950F664
+HASH = BFB21D835C4D5E08A55A2CF378FC475
